requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain-core==0.1.10
langchain-openai==0.0.5
mcp==0.1.8
langchain-mcp==0.1.2
pydantic==2.4.2
python-dotenv==1.0.0
langchain_community>=0.0.28,<0.1
